hw4

+++ src/test/java/com/epam/hw2/TrafficLightTest.java
@@ -7,15 +7,15 @@ public class TrafficLightTest {
 
     @Test
     public void testKnowLight() {
-        Assert.assertEquals((long)1, (long)TrafficLight.getLight(1));
-        Assert.assertEquals((long)1, (long)TrafficLight.getLight(2));
-        Assert.assertEquals((long)2, (long)TrafficLight.getLight(3));
-        Assert.assertEquals((long)2, (long)TrafficLight.getLight(4));
-        Assert.assertEquals((long)2, (long)TrafficLight.getLight(5));
-        Assert.assertEquals((long)3, (long)TrafficLight.getLight(6));
-        Assert.assertEquals((long)3, (long)TrafficLight.getLight(7));
-        Assert.assertEquals((long)3, (long)TrafficLight.getLight(8));
-        Assert.assertEquals((long)3, (long)TrafficLight.getLight(9));
-        Assert.assertEquals((long)3, (long)TrafficLight.getLight(0));
+        Assert.assertEquals(1, TrafficLight.getLight(1));
+        Assert.assertEquals(1, TrafficLight.getLight(2));
+        Assert.assertEquals(2, TrafficLight.getLight(3));
+        Assert.assertEquals(2, TrafficLight.getLight(4));
+        Assert.assertEquals(2, TrafficLight.getLight(5));
+        Assert.assertEquals(3, TrafficLight.getLight(6));
+        Assert.assertEquals(3, TrafficLight.getLight(7));
+        Assert.assertEquals(3, TrafficLight.getLight(8));
+        Assert.assertEquals(3, TrafficLight.getLight(9));
+        Assert.assertEquals(3, TrafficLight.getLight(0));
     }
 }
